requirements.txt
django~=5.1.3
djangorestframework~=3.15.2
psycopg2-binary

django-filter~=24.3